Mudanca na logica do jogo python e primeiro teste
