+++ jogo/bots/rl_bot.py
@@ -62,3 +62,8 @@ class RLBot:
         with open(qfile, "w") as f:
             json.dump(to_save, f, indent=2)
         print(f"[RLBot] Q-table salva em {qfile}")
+
+    def choose_card(self, player_deck, stat):
+        # escolhe a carta com maior valor no stat escolhido
+        best_card = max(self.deck, key=lambda c: (1/c[stat] if stat in ["weight","0-100"] else c[stat]))
+        return best_card
