+++ jogo/utils.py
